LayoutTests/media/modern-media-controls/volume-support/volume-support-drag-expected.txt

Testing the VolumeSupport behavior when dragging the scrubber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


By default, the volume slider should be set to 0 because the media is muted
PASS mediaController.controls.volumeSlider.value is 0
PASS muteButton.getBoundingClientRect().width became different from 0
PASS volumeSlider.getBoundingClientRect().width became different from 0

Before starting to drag the volume slider, the media should be muted still
PASS media.muted is true

We initiated a volume slider drag, the media should no longer be muted and the volume set
PASS media.muted became false
PASS media.volume became 0.5

We finished dragging to set the volume to 1
PASS media.volume became 1
PASS mediaController.controls.volumeSlider.parent.visible is true
PASS successfullyParsed is true

TEST COMPLETE
